requirements.txt
numpy==1.19.2
Flask==1.1.2
tensorflow-cpu==2.6.0
Pillow==8.3.2
h5py==3.1.0
gunicorn==20.0.4
opencv-python-headless
